ST-294: First phase package refactor, compiling
