+++ app/views/IncomeSourceSummary.scala.html
@@ -108,7 +108,7 @@
                 <p  aria-hidden="true">
                     @Link.toInternalPage(
                         id=Some("updateIncome"),
-                        url=controllers.routes.IncomeUpdateCalculatorController.howToUpdatePage(model.empId).url,
+                        url=controllers.income.estimatedPay.update.routes.IncomeUpdateCalculatorController.howToUpdatePage(model.empId).url,
                         value=Some(messages("tai.income.details.updateTaxableIncome.update"))).toHtml
                 </p>
                 <p class="visuallyhidden">
