+++ app/views/incomes/bonusPayments.scala.html
@@ -44,7 +44,7 @@ templateRenderer: uk.gov.hmrc.renderer.TemplateRenderer, partialRetriever: uk.go
     employerName = Some(employerName),
     gaCustomTitle = None
     ) {
-    @form(action = controllers.routes.IncomeUpdateCalculatorController.handleBonusPayments, Symbol("data-dynamic-form") -> "true") {
+    @form(action = controllers.income.estimatedPay.update.routes.IncomeUpdateCalculatorController.handleBonusPayments, Symbol("data-dynamic-form") -> "true") {
 
         @{inputRadioGroupStyling(
             field=bonusPaymentsForm("bonusPayments"),
