+++ app/views/incomes/estimatedPay.scala.html
@@ -81,7 +81,7 @@ messages: Messages, user: controllers.auth.TaiUser, templateRenderer: uk.gov.hmr
 
         @includes.link(
             copy=messages("tai.estimatedPay.checkTaxEstimate"),
-            url = routes.IncomeUpdateCalculatorController.handleCalculationResult.toString,
+            url = income.estimatedPay.update.routes.IncomeUpdateCalculatorController.handleCalculationResult.toString,
             isButton=true
         )
