+++ app/views/incomes/yourIncomeCalculation.scala.html
@@ -175,12 +175,12 @@
                         @if(model.isPension){
                             <p id="pensionUpdateLink">@Html(Messages(
                                 "tai.income.calculation.update.pension",
-                                Link.toInternalPage(url=routes.IncomeUpdateCalculatorController.howToUpdatePage(model.empId).toString, value=Some(Messages("tai.income.calculation.updateLink.regular"))).toHtml
+                                Link.toInternalPage(url=controllers.income.estimatedPay.update.routes.IncomeUpdateCalculatorController.howToUpdatePage(model.empId).toString, value=Some(Messages("tai.income.calculation.updateLink.regular"))).toHtml
                                 ))</p>
                         } else {
                             <p id="regularUpdateLink">@Html(Messages(
                                 "tai.income.calculation.update.regular",
-                                Link.toInternalPage(url=routes.IncomeUpdateCalculatorController.howToUpdatePage(model.empId).toString, value=Some(Messages("tai.income.calculation.updateLink.regular"))).toHtml
+                                Link.toInternalPage(url=controllers.income.estimatedPay.update.routes.IncomeUpdateCalculatorController.howToUpdatePage(model.empId).toString, value=Some(Messages("tai.income.calculation.updateLink.regular"))).toHtml
                                 ))
                             </p>
                         }
